fix: schedule scroll behavior
