+++ packages/web/src/pages/schedule.tsx
@@ -2,18 +2,10 @@ import React from 'react';
 import { NextPage } from 'next';
 import axios from 'axios';
 import dayjs from 'dayjs';
-import { useToast, UseToastOptions } from '@chakra-ui/react';
 import { Event, SerializedEvent } from '@hangar/shared';
 import { PageContainer } from '../components/layout/PageContainer';
 import { EventsList } from '../components/EventsList';
-
-const FAILED_FETCH_TOAST: UseToastOptions = {
-  title: 'An error occurred.',
-  description: 'Unable to fetch events.',
-  status: 'error',
-  duration: 3000,
-  isClosable: true,
-};
+import { openErrorToast } from '../components/utils/CustomToast';
 
 const fetchEvents: () => Promise<Event[]> = async () =>
   (await axios.get<SerializedEvent[]>('/api/event')).data.map((serializedEvent) => {
@@ -29,19 +21,21 @@ const fetchEvents: () => Promise<Event[]> = async () =>
 
 const Schedule: NextPage = () => {
   const [events, setEvents] = React.useState<Event[]>([]);
-  const toast = useToast();
 
   React.useEffect(() => {
     const fetchAndSetEvents = async () => {
       try {
         setEvents(await fetchEvents());
       } catch {
-        toast(FAILED_FETCH_TOAST);
+        openErrorToast({
+          title: 'Failed to retrieve events',
+          description: 'An unexpected error occurred',
+        });
       }
     };
 
     void fetchAndSetEvents();
-  }, [toast]);
+  }, []);
 
   return (
     <PageContainer pageTitle={'Schedule'} heading={'Schedule'}>
